Reduce type casting by using pattern matching
By using `case let` instead of `is`, we remove the additional casting when setting the component on the view.

Each case now sets a value to `view` instead of returning, this has the potential to make debugging easier as you can set a breakpoint at the very end to check the output value.

Because the framework uses constraints, we want to always set `translatesAutoresizingMaskIntoConstraints` to `false`. This is now handled before the return instead in each case.